feat: skill hooks, mob poison DoT, element WZ #14
- Rename skill Lua hooks to on_activated_<skillId> (skill_hook.go)
- Mob poison tick via MapActor timer; debuff value holds pDam; clip HP to min 1 on tick
- element_amp_from_class / element_weak_multiplier in script_common; WZ Skill.elemAttr and Mob.elemResist
- Attack, damage, magic, summon handlers; protocol attack/mob/damaged; incoming_hit constant
- Docs: poison_mist_2111003.md design notes

+++ game/constant/constant_stat.go
@@ -40,6 +40,6 @@ const (
 // Stat limits
 const (
 	STAT_MAX_STR_DEX_INT_LUK uint16 = 999   // Maximum value for STR, DEX, INT, LUK
-	STAT_MAX_HP_MP           uint16 = 30000 // Maximum value for HP and MP
+	STAT_MAX_HP_MP           uint32 = 30000
 	HP_AP_USED_MAX           uint16 = 10000 // Maximum HP/MP AP usage count
 )
