+++ game/server/handler_damage_summon.go
@@ -46,12 +46,12 @@ func (h *DamageSummon) Handle(ctx *core.ClientContext, req *request.DamageSummon
 
 	damage := req.Damage
 	if damage > 0 {
-		if summon.Hp <= uint16(damage) {
+		if summon.Hp <= damage {
 			summon.Hp = 0
 			character.RemoveSummon(summon)
 			character.Listener.OnSummonRemove(character, summon, true)
 		} else {
-			summon.Hp -= uint16(damage)
+			summon.Hp -= damage
 			character.Listener.OnSummonDamaged(character, summon, req.Unknown, damage, req.MonsterIdFrom)
 		}
 	}
